requirements.txt
aiogram==3.8.0
aiosqlite==0.20.0
httpx==0.27.0
pydantic==2.7.1
apscheduler==3.10.4
python-dotenv==1.0.1
rich==13.7.1
